requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
